Add day 06 (Wait For It - charging cars -> fast cars)
Very simple

+++ aoc23/java/src/main/java/tech/dobler/aoc23/Day_multiLineInput.java
@@ -4,11 +4,11 @@ import java.util.stream.Stream;
 
 public class Day_multiLineInput {
 
-    public int part1(Stream<String> input) {
+    public long part1(Stream<String> input) {
         return -1;
     }
 
-    public int part2(Stream<String> input) {
+    public long part2(Stream<String> input) {
         return -1;
     }
 }
